requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib
python-jose
psycopg2-binary
pydantic[email]
python-multipart